feat: merge scripts and work on account classes
