+++ src/main.js
@@ -1,18 +1,24 @@
 import account from "./data";
-import {DeleteAccountList, PasteAccountList} from "./script/account-list";
-import {LogList} from "./script/log-list";
 
-// Delete account list
+const accountBalance = document.getElementById("balance");
+const themeSwitchButton = document.getElementById("dark-mode-toggle");
+const accountName = document.getElementById("account-name");
+const accountNumber = document.getElementById("account-number");
+const enterpriseTab = document.getElementById("tab-enterprise");
+const offshoreTab = document.getElementById("tab-offshore");
+
+// Account lists
 const deleteAccountList = document.getElementById("account-delete-list");
 const deleteAccountTemplate = document.getElementById("account-delete-template");
-
-// Paste account list
 const pasteAccountList = document.getElementById("account-paste-list");
 const pasteAccountTemplate = document.getElementById("account-paste-template");
 
 // Log lists
 const accountLogList = document.getElementById("account-log-list");
 const accountOperationLogList = document.getElementById("account-operation-log-list");
+const enterpriseAccountLogList = document.getElementById("enterprise-account-log-list");
+const offshoreAccountLogList = document.getElementById("offshore-account-log-list");
+const logTemplate = document.getElementById("log-template");
 
 // Account elements
 const accountAddForm = document.getElementById("account-add-form");
@@ -47,14 +53,27 @@ const enterpriseDepositMessage = document.getElementById("enterprise-deposit-mes
 const enterpriseWithdrawForm = document.getElementById("enterprise-withdraw-form");
 const enterpriseWithdrawInput = document.getElementById("enterprise-withdraw-input");
 const enterpriseWithdrawMessage = document.getElementById("enterprise-withdraw-message");
+const enterpriseAccountName = document.getElementById("enterprise-account-name");
+const enterpriseAccountBalance = document.getElementById("enterprise-account-balance");
 
 // Offshore elements
 const offshoreDepositForm = document.getElementById("offshore-deposit-form");
 const offshoreDepositInput = document.getElementById("offshore-deposit-input");
 const offshoreDepositMessage = document.getElementById("offshore-deposit-message");
+const offshoreAccountName = document.getElementById("offshore-account-name");
+const offshoreAccountBalance = document.getElementById("offshore-account-balance");
+
+const accountAddButton = document.getElementById("account-add-button");
+const savedAccounts = document.getElementById("saved-accounts");
+
+const LOGS_PER_PAGES = 10;
+const ACCOUNT_LIMIT = 5;
 
 const appElement = document.getElementById("app");
-document.documentElement.classList.add("fleeca");
+
+/*------------------------------------*\
+  Formatters
+\*------------------------------------*/
 
 const currencyFormatter = new Intl.NumberFormat("en-US", {
   style: "currency",
@@ -67,6 +86,10 @@ const dateTimeFormatter = new Intl.DateTimeFormat("fr-FR", {
   dateStyle: "long",
   timeStyle: "short",
 });
+
+/*------------------------------------*\
+  Favorite accounts
+\*------------------------------------*/
 
 function createDeleteAccountElement(account) {
   const accountFragment = deleteAccountTemplate.content.cloneNode(true);
@@ -97,51 +120,325 @@ function createPasteAccountElement(account) {
   return accountFragment;
 }
 
-new DeleteAccountList(
-  account.favoriteAccounts,
-  deleteAccountList,
-  createDeleteAccountElement
-);
+class AccountList {
+  constructor(
+    accounts,
+    list,
+    createAccountElement
+  ) {
+    this.accounts = accounts;
+    this.list = list;
+    this.createAccountElement = createAccountElement;
+    this.addAccountsToList();
+    document.addEventListener("add-favorite-account", (event) => {
+      const account = event.detail.account;
+      this.addAccount(account, true);
+    });
+    document.addEventListener("delete-favorite-account", (event) => {
+      const { account, accountElement } = event.detail;
+      this.deleteAccount(account, accountElement);
+    });
+  }
 
-new PasteAccountList(
-  account.favoriteAccounts,
-  pasteAccountList,
-  createPasteAccountElement
-);
+  addAccountsToList() {
+    this.accounts.forEach((account) => {
+      const accountElement = this.createAccountElement(account);
+      this.list.appendChild(accountElement);
+    });
+  }
 
-new LogList(
-  account.logs,
-  accountLogList,
-  currencyFormatter,
-  dateTimeFormatter,
-  "add-account-log"
-);
+  addAccount(account) {
+    if (this.accounts.length < ACCOUNT_LIMIT) {
+      this.accounts = [account, ...this.accounts];
+      this.resetList();
+    } else {
+      throw new Error("Account limit is reached");
+    }
+  }
 
-new LogList(
-  account.logs.filter((log) => log.type === "operation"),
-  accountOperationLogList,
-  currencyFormatter,
-  dateTimeFormatter,
-  "add-account-operation-log"
-);
+  deleteAccount(account, accountElement) {
+    this.accounts = this.accounts.filter((savedAccount) => {
+      return savedAccount.number !== account.number;
+    });
+    this.resetList();
+  }
 
-/**
- * Get the current date in ISO format.
- *
- * @return {string} - The current ISO date.
- */
+  resetList() {
+    this.list.innerHTML = "";
+    this.addAccountsToList();
+  }
+
+}
+
+class DeleteAccountList extends AccountList {
+  constructor(accounts, list, createAccountElement) {
+    super(accounts, list, createAccountElement);
+    this.handleAddButton();
+  }
+
+  addAccount(account, position) {
+    super.addAccount(account, position);
+    this.handleAddButton();
+  }
+
+  deleteAccount(account, accountElement) {
+    super.deleteAccount(account, accountElement);
+    this.handleAddButton();
+  }
+
+  handleAddButton() {
+    const numberOfAccounts = this.accounts.length;
+    savedAccounts.textContent = `[${String(numberOfAccounts)}/5]`;
+    if (numberOfAccounts >= 5) {
+      if (accountAddButton.classList.contains("button--primary")) {
+        accountAddButton.classList.replace("button--primary", "button--disabled");
+        accountAddButton.setAttribute("disabled", "");
+      }
+    } else if (accountAddButton.classList.contains("button--disabled")) {
+      accountAddButton.classList.replace("button--disabled", "button--primary");
+      accountAddButton.removeAttribute("disabled");
+    }
+  }
+}
+
+class PasteAccountList extends AccountList {
+  constructor(accounts, list, createAccountElement) {
+    super(accounts, list, createAccountElement);
+  }
+}
+
+/*------------------------------------*\
+  Logs
+\*------------------------------------*/
+
+class LogList {
+  constructor(
+    logs,
+    list,
+    eventType,
+  ) {
+    this.logList = list;
+    this.logs = logs;
+    this.page = 1;
+    this.observer = new IntersectionObserver((entries) => {
+      this.getNextPage(entries);
+    });
+    this.observedLog = false;
+    this.eventType = eventType;
+    this.getPageNumber();
+    this.getPageLogs();
+    this.createPageLogs();
+    document.addEventListener(this.eventType, (event) => {
+      const log = event.detail.log;
+      this.addLog(log);
+    });
+  }
+
+  addLog(log) {
+    this.logs.unshift(log);
+    this.reset();
+  }
+
+  createLog(
+    log,
+    prepend = false,
+    observe = false
+  ) {
+    const logFragment = logTemplate.content.cloneNode(true);
+    const logRow = logFragment.querySelector(".log");
+    const logIcon = logRow.querySelector(".log__icon");
+    const logType = logRow.querySelector(".log__type");
+    const logDate = logRow.querySelector(".log__date");
+    const logAmount = logRow.querySelector(".log__amount");
+    const logReference = logRow.querySelector(".log__reference");
+    logIcon.classList.add(`log__icon--${log.type}`);
+    logAmount.classList.add(`log__amount--${log.amount > 0 ? "up" : "down"}`);
+    logType.textContent = log.label;
+    logDate.textContent = dateTimeFormatter.format(new Date(log.date));
+    logAmount.textContent = currencyFormatter.format(log.amount);
+    logReference.textContent = log.reference;
+    if (prepend) {
+      this.logList.prepend(logFragment);
+    } else {
+      this.logList.appendChild(logFragment);
+    }
+    if (observe) {
+      this.observedLog = logRow;
+      this.observer.observe(logRow);
+    }
+  }
+
+  createPageLogs() {
+    if (this.observedLog) {
+      this.observer.unobserve(this.observedLog);
+      this.observedLog = false;
+    }
+    this.pageLogs.forEach((pageLog, pageLogIndex) => {
+      const isLastLog = pageLogIndex === (this.pageLogs.length - 1);
+      const isNotLastPage = this.page < this.pageNumber;
+      if (isLastLog && isNotLastPage) {
+        this.createLog(pageLog, false, true);
+      } else {
+        this.createLog(pageLog);
+      }
+    });
+  }
+
+  getNextPage(entries) {
+    if (entries[0].isIntersecting) {
+      this.page += 1;
+      this.getPageLogs();
+      this.createPageLogs();
+    }
+  }
+
+  getPageNumber() {
+    const logNumber = this.logs.length;
+    this.pageNumber = Math.ceil( logNumber / LOGS_PER_PAGES);
+  }
+
+  getPageLogs() {
+    const firstIndex = LOGS_PER_PAGES * (this.page - 1);
+    const lastIndex = LOGS_PER_PAGES * this.page;
+    this.pageLogs = this.logs.slice(firstIndex, lastIndex);
+  }
+
+  reset() {
+    this.logList.innerHTML = "";
+    this.page = 1;
+    this.getPageLogs();
+    this.createPageLogs();
+  }
+}
+
+/*------------------------------------*\
+  Account
+\*------------------------------------*/
+
+class Account {
+  constructor(account) {
+    this.name = account.name;
+    this.number = account.number;
+    this.balance = account.balance;
+    this.favoriteAccounts = account.favoriteAccounts;
+    this.logs = account.logs;
+    this.theme = account.theme;
+    if (this.theme === "dark") {
+      themeSwitchButton.checked = true;
+      document.documentElement.classList.add("dark");
+    }
+    this.displayAccount();
+    this.displayFavoriteAccounts();
+    this.displayAccountLogs();
+    this.hasEnterprise = account.hasEnterprise;
+    this.hasOffshore = account.hasOffshore;
+    if (this.hasEnterprise) {
+      enterpriseTab.classList.remove("tab-list__item--hidden");
+      this.enterpriseOwner = account.enterprise.owner;
+      this.enterpriseBalance = account.enterprise.balance;
+      this.enterpriseLogs = account.enterprise.logs;
+      this.displayEnterpriseAccount();
+      this.displayEnterpriseAccountLogs();
+    }
+    if (this.hasOffshore) {
+      offshoreTab.classList.add("paper__button--visible");
+      this.offshoreOwner = account.offshore.owner;
+      this.offshoreBalance = account.offshore.balance;
+      this.offshoreLogs = account.offshore.logs;
+      this.displayOffshoreAccount();
+      this.displayOffshoreAccountLogs();
+    }
+    themeSwitchButton.addEventListener("change", (event) => {
+      this.handleTheme(event);
+    });
+    document.addEventListener("update-account-balance", (event) => {
+      const amount = event.detail.amount;
+      this.balance += amount;
+      this.displayBalance(this.balance, accountBalance);
+    });
+    document.addEventListener("update-enterprise-account-balance", (event) => {
+      const amount = event.detail.amount;
+      this.enterpriseBalance += amount;
+      this.displayBalance(this.enterpriseBalance, enterpriseAccountBalance);
+    });
+    document.addEventListener("update-offshore-account-balance", (event) => {
+      const amount = event.detail.amount;
+      this.offshoreBalance += amount;
+      this.displayBalance(this.offshoreBalance, offshoreAccountBalance);
+    });
+  }
+
+  handleTheme(event) {
+    this.theme = event.target.checked ? "dark" : "light";
+    document.dispatchEvent(new CustomEvent("theme-update", { theme: this.theme }));
+    if (this.theme === "dark") {
+      if (!document.documentElement.classList.contains("dark")) {
+        document.documentElement.classList.add("dark");
+      }
+    } else {
+      if (document.documentElement.classList.contains("dark")) {
+        document.documentElement.classList.remove("dark");
+      }
+    }
+  }
+
+  displayBalance(balance, balanceElement) {
+    if (balance < 1) {
+      if (!balanceElement.classList.contains("balance__value--negative")) {
+        balanceElement.classList.add("balance__value--negative");
+      }
+    } else {
+      if (balanceElement.classList.contains("balance__value--negative")) {
+        balanceElement.classList.remove("balance__value--negative");
+      }
+    }
+    balanceElement.textContent = currencyFormatter.format(balance);
+  }
+
+  displayAccount() {
+    accountName.textContent = this.name;
+    accountNumber.textContent = this.number;
+    this.displayBalance(this.balance, accountBalance);
+  }
+
+  displayEnterpriseAccount() {
+    enterpriseAccountName.textContent = this.enterpriseOwner;
+    this.displayBalance(this.enterpriseBalance, enterpriseAccountBalance);
+  }
+
+  displayOffshoreAccount() {
+    offshoreAccountName.textContent = this.offshoreOwner;
+    this.displayBalance(this.offshoreBalance, offshoreAccountBalance);
+  }
+
+  displayFavoriteAccounts() {
+    new DeleteAccountList(this.favoriteAccounts, deleteAccountList, createDeleteAccountElement);
+    new PasteAccountList(this.favoriteAccounts, pasteAccountList, createPasteAccountElement);
+  }
+
+  displayAccountLogs() {
+    new LogList(this.logs, accountLogList, "add-account-log");
+    new LogList(this.logs.filter(log => log.type === "operation"), accountOperationLogList, "add-account-operation-log");
+  }
+
+  displayEnterpriseAccountLogs() {
+    new LogList(this.enterpriseLogs, enterpriseAccountLogList, "add-enterprise-account-log");
+  }
+
+  displayOffshoreAccountLogs() {
+    new LogList(this.offshoreLogs, offshoreAccountLogList, "add-offshore-account-log");
+  }
+}
+
+/*------------------------------------*\
+  Forms
+\*------------------------------------*/
+
 function getCurrentDate() {
   const date = new Date();
   return date.toISOString();
 }
 
-/**
- * Check account name input.
- *
- * @param input
- * @param message
- * @returns {boolean}
- */
 function checkAccountNameField(input, message) {
   if (!input.validity.valid) {
     if (input.validity.valueMissing) {
@@ -157,13 +454,6 @@ function checkAccountNameField(input, message) {
   return true;
 }
 
-/**
- * Check account number input.
- *
- * @param input
- * @param message
- * @returns {boolean}
- */
 function checkAccountNumberField(input, message) {
   if (!input.validity.valid) {
     if (input.validity.valueMissing) {
@@ -181,13 +471,6 @@ function checkAccountNumberField(input, message) {
   return true;
 }
 
-/**
- * Check amount input.
- *
- * @param input
- * @param message
- * @returns {boolean}
- */
 function checkAmountField(input, message) {
   if (!input.validity.valid) {
     if (input.validity.valueMissing) {
@@ -206,13 +489,6 @@ function checkAmountField(input, message) {
   return true;
 }
 
-/**
- * Check withdraw amount input.
- *
- * @param input
- * @param message
- * @returns {boolean}
- */
 function checkWithdrawAmount(input, message) {
   const amount = Number(input.value);
   if (account.balance < amount) {
@@ -222,13 +498,6 @@ function checkWithdrawAmount(input, message) {
   return true;
 }
 
-/**
- * Check reference input.
- *
- * @param input
- * @param message
- * @returns {boolean}
- */
 function checkReferenceField(input, message) {
   if (!input.validity.valid) {
     if (input.validity.valueMissing) {
@@ -244,11 +513,6 @@ function checkReferenceField(input, message) {
   return true;
 }
 
-/**
- * Handle account add form.
- *
- * @param event - The event.
- */
 function handleAccountAddForm(event) {
   event.preventDefault();
   const accountNameFieldIsValid = checkAccountNameField(accountAddNameInput, accountAddNameMessage);
@@ -262,11 +526,6 @@ function handleAccountAddForm(event) {
   }
 }
 
-/**
- * Handle deposit form.
- *
- * @param event - The event.
- */
 function handleDepositForm(event) {
   event.preventDefault();
   const amountFieldIsValid = checkAmountField(depositInput, depositMessage);
@@ -286,11 +545,6 @@ function handleDepositForm(event) {
   }
 }
 
-/**
- * Handle withdraw form.
- *
- * @param event - The event.
- */
 function handleWithdrawForm(event) {
   event.preventDefault();
   const withdrawFieldIsValid = checkAmountField(withdrawInput, withdrawMessage);
@@ -311,11 +565,6 @@ function handleWithdrawForm(event) {
   }
 }
 
-/**
- * Handle transfer form.
- *
- * @param event - The event.
- */
 function handleTransferForm(event) {
   event.preventDefault();
   const transferAmountFieldIsValid = checkAmountField(transferAmountInput, transferAmountMessage);
@@ -337,11 +586,6 @@ function handleTransferForm(event) {
   }
 }
 
-/**
- * Handle enterprise deposit form.
- *
- * @param event - The event.
- */
 function handleEnterpriseDepositForm(event) {
   event.preventDefault();
   const amountFieldIsValid = checkAmountField(enterpriseDepositInput, enterpriseDepositMessage);
@@ -360,11 +604,6 @@ function handleEnterpriseDepositForm(event) {
   }
 }
 
-/**
- * Handle enterprise withdraw form.
- *
- * @param event - The event.
- */
 function handleEnterpriseWithdrawForm(event) {
   event.preventDefault();
   const withdrawFieldIsValid = checkAmountField(enterpriseWithdrawInput, enterpriseWithdrawMessage);
@@ -384,11 +623,6 @@ function handleEnterpriseWithdrawForm(event) {
   }
 }
 
-/**
- * Handle offshore deposit form.
- *
- * @param event - The event.
- */
 function handleOffshoreDepositForm(event) {
   event.preventDefault();
   const amountFieldIsValid = checkAmountField(offshoreDepositInput, offshoreDepositMessage);
@@ -415,6 +649,14 @@ enterpriseDepositForm.addEventListener("submit", handleEnterpriseDepositForm);
 enterpriseWithdrawForm.addEventListener("submit", handleEnterpriseWithdrawForm);
 offshoreDepositForm.addEventListener("submit", handleOffshoreDepositForm);
 
+/*------------------------------------*\
+  Init
+\*------------------------------------*/
+
 setTimeout(() => {
   appElement.classList.replace("app--loading", "app--loaded");
 }, 1000);
+
+new Account(account);
+
+document.documentElement.classList.add("fleeca");
